+++ mycodo/inputs/atlas_ec.py
@@ -22,6 +22,7 @@ INPUT_INFORMATION = {
         'ftdi_location',
         'i2c_location',
         'uart_location',
+        'uart_baud_rate',
         'period',
         'pre_output',
         'log_level_debug'
@@ -35,7 +36,9 @@ INPUT_INFORMATION = {
     'interfaces': ['I2C', 'UART', 'FTDI'],
     'i2c_location': ['0x66'],
     'i2c_address_editable': True,
-    'uart_location': '/dev/ttyAMA0'
+    'ftdi_location': '/dev/ttyUSB0',
+    'uart_location': '/dev/ttyAMA0',
+    'uart_baud_rate': 9600,
 }