+++ mycodo/inputs/atlas_orp.py
@@ -58,6 +58,7 @@ INPUT_INFORMATION = {
     'interfaces': ['I2C', 'UART', 'FTDI'],
     'i2c_location': ['0x66'],
     'i2c_address_editable': True,
+    'ftdi_location': '/dev/ttyUSB0',
     'uart_location': '/dev/ttyAMA0',
     'uart_baud_rate': 9600,